% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Problem.R
\name{adapt}
\alias{adapt}
\title{Adapt an existing design}
\usage{
adapt(design, prior, partial, verbosity = 3, timelimit = 180, ...)
}
\arguments{
\item{design}{asd}

\item{prior}{ad}

\item{partial}{asd}

\item{verbosity}{asd}

\item{timelimit}{maximum time spent in ILP solver (in seconds)}

\item{...}{optional arguments}
}
\description{
Adapt an existing design
}
\examples{
prior   <- Beta(5, 7)
problem <- Problem(
    minimise(SampleSize(prior)),
    Power(prior \%|\% 0.2) <= 0.05,
    Power(prior  >= 0.3) >= 0.80
)
design <- optimise_design(problem)

}
