% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{binomial_test}
\alias{binomial_test}
\title{Single-stage binomial test}
\usage{
binomial_test(alpha, beta, p0, p1)
}
\arguments{
\item{alpha}{maximal type one error rate}

\item{beta}{maximal type two error rate}

\item{p0}{boundary of the null hypothesis}

\item{p1}{point alternative}
}
\description{
Calculate sample size and critical value for a single-stage binomial test.
}
