% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\docType{methods}
\name{condition}
\alias{condition}
\alias{condition,Prior-method}
\alias{\%|\%}
\alias{\%|\%,Prior,numeric-method}
\title{Conditioning of prior distributions}
\usage{
condition(prior, low = 0, high = 1)

\S4method{condition}{Prior}(prior, low = 0, high = 1)

prior \%|\% interval

\S4method{\%|\%}{Prior,numeric}(prior, interval)
}
\arguments{
\item{prior}{\code{\link{Prior}} distribution object representing a prior on
the unit interval [0,1]}

\item{low}{[optional] lower boundary to condition on}

\item{high}{[optional] upper boundary to condition on}
}
\description{
\code{\link{Prior}} objects can be conditioned on lower or upper boundaries,
intervals, or single values to restrict their support.
Various shorthand notations are supported (see examples).
}
\examples{
condition(Beta(2, 3), low = 0.2)
condition(Beta(2, 3), high = 0.9)
condition(Beta(2, 3), low = 0.2, high = 0.9)

Beta(5, 7) \%|\% c(0.2, 0.7) # shorthand for conditioning on interval
Beta(5, 7) \%|\% 0.5 # condition on single point

Beta(5, 7) <= 0.5 # = condition(Beta(5, 7), high = 0.5)
0.2 <= Beta(5, 7) # = condition(Beta(5, 7), low = 0.2)

}
\seealso{
\code{\link{Prior}}, \code{\link{updating}} for Bayesian updating
}
