% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{constraints}
\alias{constraints}
\alias{>=,Power,numeric-method}
\title{Power and type-one-error-rate constraints}
\usage{
\S4method{>=}{Power,numeric}(e1, e2)
}
\arguments{
\item{score}{either a score of type \code{Power} or \code{TypeOneErrorRate},
see \code{\link{scores}}.}

\item{threshold}{lower (Power) or upper (type one error rate) threshold for
the constraint}

\item{conditional}{vector of length two giving the corridor for conditional
power/type one error; defaults to \code{c(.5, .99)} for power and
\code{c(.001, .99)} for type one error rate. Set to c(0, 1)
to deactivate completely.}
}
\description{
Power and type-one-error-rate constraints
}
\examples{
\donttest{
  load_julia_package()
  # maximal type one error rate constraint
  toer_cnstr  <- Power(Beta(5, 7) \%|\% 0.2) <= 0.05
  # power constraint with conditional power (given x1) restricted to 25\%-100\%
  power_cnstr <- (Power(Beta(5, 7) >=  0.3) >= 0.8) \%>\%
    conditional(c(0.25, 1.00))
}
}
\seealso{
\code{\link{scores}}, \code{\link{objectives}}
}
