% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R, R/plot_designs.R
\docType{class}
\name{Design}
\alias{Design}
\alias{sample_size}
\alias{critical_value}
\alias{futility_region}
\alias{efficacy_region}
\alias{continuation_region}
\alias{pmf}
\alias{probability_to_reject}
\alias{expected_sample_size}
\alias{sample_space}
\alias{reject}
\alias{as_tibble.Design}
\alias{plot.Design}
\alias{plot_designs}
\title{Generic two-stage designs for single-arm trials with binary endpoint}
\usage{
Design(n2, c2, label = "")

sample_size(design, x1 = NULL)

critical_value(design, x1)

futility_region(design)

efficacy_region(design)

continuation_region(design)

pmf(design, x1, x2, p)

probability_to_reject(design, p, x1 = NULL)

expected_sample_size(design, p)

sample_space(design)

reject(x1, x2, design)

\method{as_tibble}{Design}(x, ...)

\method{plot}{Design}(x, tbl_power_annotations = NULL, ...)

plot_designs(..., tbl_power_annotations = NULL, min.segment.length = Inf)
}
\arguments{
\item{n2}{numeric vector with stage-two sample sizes for x1 = 0, 1, ..., n1}

\item{c2}{numeric vectot with stage-two decision boundary for x1 = 0, 1, ..., n1;
the design rejects H0: p <= p0 iff X2 > c_2(X1); use +/- Inf for early futility/efficacy (n2 = 0).}

\item{label}{an optional label to use for pretty plotting etc.}

\item{design}{a \code{Design} object}

\item{x1}{stage-one number of results}

\item{x2}{stage-two number of responses}

\item{p}{response probability}

\item{...}{one ore more \code{Design} objects to plot}

\item{tbl_power_annotations}{a \code{\link{tibble}} with annotation data for
plotted in the expected sample size and power plots.}

\item{min.segment.length}{passed on to ggrepel}
}
\description{
[todo]
}
