% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectives.R
\docType{class}
\name{objectives}
\alias{objectives}
\alias{minimise_maximal_sample_size}
\alias{minimise}
\alias{maximise}
\title{Objective functions for optimal two-stage designs}
\usage{
minimise_maximal_sample_size(lambda, prior)

minimise(score)

maximise(score)
}
\arguments{
\item{lambda}{minimax sample size alone does not always lead to a unique
minimiser. Instead a combination of \code{(1 - lambda)} times the maximal
sample size and \code{lambda} times the expected sample size under the
given \code{prior} is minimised.}

\item{prior}{\code{\link{Prior}} object}

\item{score}{\code{Score} object to be minimised/maximised}
}
\description{
Objective functions for optimal two-stage designs
}
\seealso{
\code{\link{Design}}, \link{constraints}, \link{scores}, \code{\link{Prior}}
}
