% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Problem.R
\docType{class}
\name{optimise-design}
\alias{optimise-design}
\alias{Problem}
\alias{optimise_design}
\title{Optimise a trial design}
\usage{
Problem(objective, toer, power, label = "", unimodal = FALSE, ...)

optimise_design(problem, verbosity = 3, timelimit = 180, ...)
}
\arguments{
\item{objective}{object of class \code{Objective} (see \link{objectives})}

\item{toer}{type one error rate constraint (see \link{constraints})}

\item{power}{power constraint (see \link{constraints})}

\item{label}{label for the problem}

\item{unimodal}{force unimodal solution (can be time consuming!)}

\item{...}{optional arguments}

\item{problem}{object of class \code{Problem}}

\item{verbosity}{verbosity level of optimisation procedure}

\item{timelimit}{maximum time spent in ILP solver (in seconds)}
}
\description{
[todo]
}
\examples{
\donttest{
  prior   <- Beta(5, 7)
  problem <- Problem(
      minimise(SampleSize(prior)),
      Power(prior \%|\% 0.2) <= 0.05,
      Power(prior  >= 0.3) >= 0.80
  )
  design <- optimise_design(problem)
}
}
\seealso{
\link{scores}, \link{constraints}, \link{objectives}, \link{Design}
}
