% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\docType{class}
\name{scores}
\alias{scores}
\alias{evaluate}
\alias{evaluate,Score,Design-method}
\title{Scores}
\usage{
evaluate(score, design, ...)

\S4method{evaluate}{Score,Design}(score, design, x1 = NULL, partial = c(0, 0))
}
\arguments{
\item{score}{\code{Score} object to evaluate}

\item{design}{\code{\link{Design}} object to evaluate score on}

\item{x1}{completely observed number of stage one responses under design; if given,
thew score is evaluated conditional on \code{x1}}

\item{partial}{partial observations \code{c(x, n)} in stage two (if \code{x1} is given) or stage
one (if \code{x1} is not given)}
}
\description{
Score objects allows the computation of conditional and unconditional
performance measures for \code{\link{Design}} objects.
The \code{evaluate} generic function and its methods allow to compute
scores conditional on observed (partial) stage one results or stage
two results or completely unconditional.
}
\seealso{
\code{\link{SampleSize}}, \code{\link{Power}}, \code{\link{TypeOneErrorRate}}
}
