% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{updating}
\alias{updating}
\alias{update}
\alias{update,PointMass-method}
\alias{update,Beta-method}
\alias{update,BetaMixture-method}
\alias{update,GenericDistribution-method}
\alias{update,JeffreysPrior-method}
\title{Bayesian updating of prior distributions}
\usage{
\S4method{update}{PointMass}(object, x, n)

\S4method{update}{Beta}(object, x, n)

\S4method{update}{BetaMixture}(object, x, n)

\S4method{update}{GenericDistribution}(object, x, n)

\S4method{update}{JeffreysPrior}(object, x, n)
}
\arguments{
\item{object}{\code{\link{Prior}} distribution object to update}

\item{x}{number of responses out of \code{n} individuals}

\item{n}{number of individuals out of which \code{x} had a response}
}
\description{
All prior distributions described in \code{\link{Prior}} can be updated with
binomial observations (\code{x} out of \code{n} responses).
}
\examples{
\dontrun{
  # point mass distributions are invariant under updating
  update(PointMass(.4), 3, 10)
}

\dontrun{
  update(Beta(5, 7), 3, 10) # same as Beta(8, 14)
}

\dontrun{
  update(1/3*Beta(5, 7) + 2/3*Beta(1,1), 3, 10) # update mixtures
}
\dontrun{
  design    <- Design(c(0, 30, 25, 0), c(Inf, 10, 7, -Inf))
  prior     <- JeffreysPrior(design)
  posterior <- update(prior, 3, 10) # results in a GenericDistribution object (no analytical update)
  update(posterior, 2, 5) # the generic posterior of a Jeffreys prior can also be updated again
}
}
\seealso{
\code{\link{Prior}}, \code{\link{condition}}
}
