% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{Prior}
\alias{Prior}
\alias{density,Prior-method}
\alias{cdf}
\alias{cdf,Prior,numeric-method}
\alias{mean,Prior-method}
\alias{PointMass}
\alias{Beta}
\alias{JeffreysPrior}
\title{Prior distributions for response probability}
\usage{
\S4method{density}{Prior}(x, p)

cdf(prior, p, ...)

\S4method{cdf}{Prior,numeric}(prior, p)

\S4method{mean}{Prior}(x)

PointMass(p)

Beta(a, b)

JeffreysPrior(design)
}
\arguments{
\item{x}{\code{Prior} distribution object (\code{density})}

\item{p}{probability atom, i.e. the response probability equals \code{p}
almost surely.}

\item{prior}{\code{Prior} distribution object}

\item{a}{Beta distribution parameter}

\item{b}{Beta distribution paramter}

\item{design}{\code{\link{Design}} object}
}
\description{
\code{badr} supports \code{PointMass}, \code{Beta}, \code{BetaMixture}, and
\code{JeffreysPrior} distributions for the reponse probability.
The probability density function and cumulative probability functions are
available via \code{density} and \code{cdf} methods.
The mean of a distribution can quickly be accessed via \code{mean}.
}
\examples{
\donttest{
  density(Beta(1, 1), seq(0, 1, .1)) == 1 # uniform distribution on [0, 1]
}
\donttest{
  cdf(PointMass(1/3), c(0.3, 1/3)) == c(0, 1)
}
\donttest{
  mean(Beta(5, 7)) == 5/(5 + 7)
}
\donttest{
  PointMass(0.4)
}
\donttest{
  Beta(5, 7)
  1/3*Beta(5, 7) + 2/3*Beta(1,1) # create a BetaMixture distribution
 }
\donttest{
  design <- Design(c(0, 30, 25, 0), c(Inf, 10, 7, -Inf))
  JeffreysPrior(design)
}
}
\seealso{
\code{\link{condition}} for restricting the support of a prior,
\code{\link{updating}} for Bayesian posterior updates
}
